Cleanup social sample
